Apaha 0.1

+++ openvpn/control.py
@@ -17,10 +17,9 @@ def adduser(username,password,group):
 		password = crypt.crypt(password, getsalt())
 		RadUserList(username=username,attribute="Crypt-Password",op=':=',value=password).save()
 		RadUserGroup(username=username,groupname=group,priority=1).save()
-
-		return 'Success'
+		return True
 	else:
-		return 'False'
+		return False
 
 def updateuser(username,password):
 	olduserlist = []
@@ -34,15 +33,18 @@ def updateuser(username,password):
 		return 'False'
 
 def deluser(username):
-	olduserlist = []
-	for i in RadUserList.objects.all().values('username'):
-		olduserlist.append(i['username'])
-	if username in olduserlist:
-		RadUserList.objects.filter(username=username).delete()
-		RadUserGroup.objects.filter(username=username).delete()
-		return 'Success'
+	if sqlfilter(username):
+		olduserlist = []
+		for i in RadUserList.objects.all().values('username'):
+			olduserlist.append(i['username'])
+		if username in olduserlist:
+			RadUserList.objects.filter(username=username).delete()
+			RadUserGroup.objects.filter(username=username).delete()
+			return True
+		else:
+			return False
 	else:
-		return 'False'
+		return False
 
 
 def addserver(serverip,serverdomain,enable):
@@ -83,15 +85,25 @@ def delserver(serverip):
 	else:
 		return 'False'
 
-def addgroupuser(group,userlist):
-	InsertList = []
-	for i in list(set(userlist)):
-		InsertList.append(RadUserGroup(username=i,groupname=group,priority=1))
-	try:
-		RadUserGroup.objects.bulk_create(InsertList)
-		return 'Success'
-	except Exception,e:
-		return e
+def addguser(userlist,password,group):
+	# InsertList = []
+	# olduserlist = []
+	#
+	# for i in RadUserList.objects.all().values('username'):
+	# 	olduserlist.append(i['username'])
+	# for username in list(set(userlist)):
+	# 	if not username == '':
+	# 		if sqlfilter(username)
+	# 			if username not in olduserlist:
+	# 				password = crypt.crypt(password, getsalt())
+	# 				InsertList.append(RadUserList(username=username,attribute="Crypt-Password",op=':=',value=password))
+	# 				InsertList.append(RadUserGroup(username=i,groupname=group,priority=1))
+	# try:
+	# 	RadUserGroup.objects.bulk_create(InsertList)
+	# 	return True
+	# except Exception:
+	# 	return False
+	print 1
 
 def status():
 	from openvpn.models import RadAcct
@@ -99,4 +111,11 @@ def status():
 	Onlinelist = RadAcct.objects.all().filter(acctstopdelay=None).values('username', 'groupname', 'nasipaddress','acctstarttime', 'callingstationid','framedipaddress')
 	return Onlinelist
 
+
+def sqlfilter(temp):
+	import re
+	if re.match(r'^[0-9a-zA-Z_]+$',temp):
+		return temp
+	else:
+		return False
 #def history():
